requirements.txt
asttokens==3.0.1
bash==0.6
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
parso==0.8.6
pefile==2024.8.26
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.20.0
PyMuPDF==1.27.2.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
pyzmq==27.1.0
reportlab==4.4.10
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.8
yara-python==4.5.4
